BehaviourTreeNode – fixed numerous bugs
I realised that caching the current state and status values was causing the node to skip states, by pre-emptively reading leftover values and moving on instead of calculating them based on the current state. Now, the values are always calculated based on the actual current state.
Additionally, rather than pre-emptively trying to calculate every single state and move all the way to the next one that is active, the system will now just go to the next state and then increment the next frame if it’s blocked or completed. This allows the system to be much simpler and with less points of failure. This also means that if the base node is fully completed, it can loop around to the start and run the entire tree's logic again without needing to switch to a different state.