FiniteStateMachine bugfix: ensuring a transition with a 'null' destination still functions
If the transition destination is null, that means the state machine transitions to the 'exit 'state' and stops functioning (unless an 'any state' transition happens). I realised that the transition could potentially fail because of a missing entry in the state list (a check is run to ensure the 'to' state is actually part of the state machine), so I ensured that check also checks if a state is present.